% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mto_predictors.R
\name{base_predictor}
\alias{base_predictor}
\title{Base predictor}
\usage{
base_predictor(data, model, excl, fix = NULL, cc = NULL, return_vars = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with data to predict probability of having GA subscription}

\item{model}{Apollo model}

\item{excl}{Coefficient vector to exclude (\code{c("c1", "c2", "scale")} will always be excluded)}

\item{fix}{Named vector. Fix coefficients to certain values}

\item{cc}{Calibration constant (cutoff parameter)}
}
\value{
\code{data.frame} with probability vectors
}
\description{
Base predictor
}
