% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_cache_export.R
\name{change_cache_folder}
\alias{change_cache_folder}
\title{Sets the cache folder or changes it if it was already set to save caches.}
\usage{
change_cache_folder(folder = null, verbose = FALSE)
}
\arguments{
\item{folder}{Folder to set as a cache folder. The default value is NULL,
which triggers the check_users_choice_if_cache function that provides some
options to the user to use it as a cache folder, a temporary one, or disable caching.}

\item{verbose}{Boolean. If TRUE, it provides information when the
cache folder is set. Otherwise, it only prints a warning when there is an error.}
}
\value{
No return value, called for side effects
}
\description{
Sets the cache folder or changes it if it was already set to save caches.
}
\examples{
change_cache_folder("my_cache_folder", verbose = TRUE)
}
