% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel}
\alias{excel}
\title{Creates an excel file from a data.frame or a list of data.frame or from
Rapi_GETPREP object.}
\usage{
excel(
  dfs = null,
  file_name = null,
  folder = null,
  .debug = FALSE,
  env = rlang::caller_env(),
  ...
)
}
\arguments{
\item{dfs}{object or list of data frame to write}

\item{file_name}{file name to save}

\item{folder}{folder to save file}

\item{.debug}{for internal use}

\item{env}{environment}

\item{...}{for future versions}
}
\value{
it returns object or list of data frame back
}
\description{
The excel() function creates an excel file according to the object given.
data.frame or List of data frame or Rapi_GETPREP object can be passed..
}
\examples{
\dontrun{
excel(data.frame(a = 1:3), file_name = "test1.xlsx", folder = ".")
}
}
