% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodeBylatlng.R
\name{geocodeBylatlng}
\alias{geocodeBylatlng}
\title{geocodeBylatlng}
\usage{
geocodeBylatlng(latlng, key)
}
\arguments{
\item{latlng}{Character class of latitude and longitude separated by ","}

\item{key}{need a vaild API key from user}
}
\value{
Character class of the address of the given latitude and longitude.
}
\description{
Returns address of the given latitude and longitude.
}
\examples{
geocodeBylatlng("37.56654, 126.978","AIzaSyCbnQWy0geP8Md3nJxfHevjKSx9TK3xI_w")
geocodeBylatlng("37.55723, 127.0453","AIzaSyCbnQWy0geP8Md3nJxfHevjKSx9TK3xI_w")
}
