% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFromCSV.R
\name{getDataFromCSV}
\alias{getDataFromCSV}
\title{Google Geocoding API connector for CSV files}
\usage{
getDataFromCSV(filePath)
}
\arguments{
\item{filePath}{A csv file path}
}
\value{
A data.frame containing the resulting \code{longitude/s}, \code{latitude/s} and \code{location1/es}
}
\description{
Takes a csv file path (with the required columns \code{location1} and/or \code{Latitude} and
 \code{Longitude}) as a parameter and creates a vector with the information that will be passed to the 
\code{getInforFromInput} function that will connect with the Google Geocoding API and get the corresponding output.
}
\examples{
#getDataFromCSV(filePath)


}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}
