% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInforFromInput.R
\name{getInforFromInput}
\alias{getInforFromInput}
\title{get Infor From User Input}
\usage{
getInforFromInput(address)
}
\arguments{
\item{address}{A character vector with of address/es and/or some coordinates (latitude and longitude)}
}
\value{
A data.frame containing the resulting \code{longitude/s}, \code{latitude/s} and \code{address/es}
}
\description{
Loops over the vector of addressInfo and/or coordinates, passes each one of them as a parameter 
to an internal function that connects to the \code{Google Geocoding API} via URL and gets the output. Then it 
returns the output of the request into a formatted dataframe with the geocoding information.
}
\examples{
getInforFromInput(c("Spain", "smålandsvägen 42D, Linghem"))



}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}
