% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_op_utils.R
\name{\%inn\%}
\alias{\%inn\%}
\title{\%inn\%}
\usage{
x \%inn\% table
}
\arguments{
\item{x}{Character to check if it exists in a vector or list.}

\item{table}{List, data frame, or any vector.}
}
\value{
Logical value TRUE if it exists, FALSE if it does not.
}
\description{
Checks if the second parameter includes the first one as a value, a column name, or a name.
}
\examples{
.check <- "a" \%inn\% data.frame(a = 1:5)
}
