% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{memory_model}
\alias{memory_model}
\title{Memory model}
\description{
Memory model
}
\examples{
\dontrun{
High memory addresses
+----------------+
|  Heap          | <- dynamic allocation, persists until freed (malloc() and free())
+----------------+
|                |
|  (unused)      |
|                |
+----------------+
|  Stack         | <- grows downward, auto allocation (e.g., callstack)
+----------------+
Low memory addresses
  }
}
