% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.Rapi_GETPREP}
\alias{print.Rapi_GETPREP}
\title{print.Rapi_GETPREP
Generic method for S3 Rapi_GETPREP object}
\usage{
\method{print}{Rapi_GETPREP}(x, ...)
}
\arguments{
\item{x}{S3 Rapi_GETPREP object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
S3 Rapi_GETPREP object
}
\description{
print.Rapi_GETPREP
Generic method for S3 Rapi_GETPREP object
}
\examples{
\dontrun{

obj <- get_series(template_test())
print(obj)
}
}
