% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_for_build_step.R
\name{Rapi-build-checks}
\alias{Rapi-build-checks}
\title{Dummy examples to check build step}
\description{
This page checks some examples on build process to avoid missing
}
\examples{

obj <- get_series(template_test(), debug = TRUE)
x <- print(obj)
if (!inherits(x, "Rapi_GETPREP")) {
  stop()
}

}
\keyword{internal}
