\name{RAPI-package}
\alias{RAPI-package}
\alias{RAPI}
\docType{package}
\title{
\packageTitle{RAPI}
}
\description{
\packageDescription{RAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RAPI}
\packageIndices{RAPI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RAPI}

Maintainer: \packageMaintainer{RAPI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
