% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAP.R
\docType{package}
\name{RAPI}
\alias{RAPI}
\alias{RAPI-package}
\title{A package for connecting to Google Geocoding API}
\description{
The \code{RAPI} package provides an interface to connect with the \code{Google Geocoding API} with 
functions that convert adresses into coordinates and vice versa.
}
\details{
The package provides the following functions:
\itemize{
   \item \code{geocodeBylatlng}: Get google address according to specific x, y  entered by user
   \item \code{getAddressInfoByLocation}: Get google x, y according to specific address and valid API key
}
}
\references{
Google Geocoding API - \url{https://developers.google.com/maps/documentation/geocoding/intro}
}
\author{
\itemize{
     \item Nour Qweder - \url{https://github.com/nourqweder}
     \item Young Lee - \url{https://github.com/nourqweder}
     \item Samia Butt - \url{https://github.com/nourqweder}
   }
}
