% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{readCSVFile}
\alias{readCSVFile}
\title{readCSVFile}
\usage{
readCSVFile(csvFile = "./Data/NYPD_7_Major_Felony_Incidents.csv")
}
\arguments{
\item{csvFile}{file to do the data preparation and load our dataset}
}
\value{
Load the Data and Prepare It for Mapping
}
\description{
Load the Data and Prepare It for Mapping
 If the data is small enough, you can prefer to do everything about the data in the
}
