% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{apollo_attach}
\alias{apollo_op}
\alias{apollo_detach}
\alias{apollo_combineModels}
\alias{apollo_avgInterDraws}
\alias{apollo_prepareProb}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{apollo}{\code{\link[apollo]{apollo_attach}}, \code{\link[apollo]{apollo_avgInterDraws}}, \code{\link[apollo]{apollo_combineModels}}, \code{\link[apollo]{apollo_detach}}, \code{\link[apollo]{apollo_op}}, \code{\link[apollo]{apollo_prepareProb}}}
}}

