% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_df_utils.R
\name{remove_columns}
\alias{remove_columns}
\title{Remove a column or columns from a data.frame.}
\usage{
remove_columns(df, column_names, verbose = FALSE)
}
\arguments{
\item{df}{Data.frame or tibble.}

\item{column_names}{Column name or column names as a character vector.}

\item{verbose}{Boolean, provides extra information when removing a column.}
}
\value{
Data.frame.
}
\description{
Remove a column or columns from a data.frame.
}
\examples{
df <- remove_columns(cars, "speed")

}
