% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_globals.R
\name{template_test}
\alias{template_test}
\title{template_test
creates a string template for testing and example purposes}
\usage{
template_test()
}
\value{
a string template that includes ID examples from different sources
}
\description{
template_test
creates a string template for testing and example purposes
}
\examples{
template_test()
}
