% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_description.R
\name{variable_description}
\alias{variable_description}
\title{Get variable descriptions}
\usage{
variable_description(
  model = c("all", "ga", "ca", "ht", "re", "cs", "bi", "wfh"),
  vars = NULL
)
}
\arguments{
\item{model}{One of \code{c("all", "ga", "ca", "ht", "re", "cs", "bi", "wfh")}}

\item{vars}{Select variables of interest. Only pass if \code{model = "all"}}
}
\value{
\code{data.frame}
}
\description{
Get variable descriptions
}
