% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_verbose.R
\name{verbose_off}
\alias{verbose_off}
\title{Turn Off Verbose Mode}
\usage{
verbose_off()
}
\value{
The function has no return value.
}
\description{
This function turns off verbose mode, suppressing additional informational
output. It is useful when you want to limit the amount of information
displayed during the execution of certain operations.
}
\details{
Verbose mode is often used to provide detailed information about the
progress of a function or operation. By calling \code{verbose_off}, you can
disable this additional output.

The \code{options("Rapi_verbose" = FALSE)} line sets the verbose option to
\code{FALSE}, silencing additional messages.
}
\examples{
verbose_off()

}
\seealso{
\code{\link{verbose_on}}: Turn on verbose mode.
}
