% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_verbose.R
\name{verbose_on}
\alias{verbose_on}
\title{Turn On Verbose Mode}
\usage{
verbose_on()
}
\value{
The function has no explicit return value.
}
\description{
This function turns on verbose mode, enabling additional informational
output. It is useful when you want to receive detailed information about
the progress of certain operations.
}
\details{
Verbose mode is designed to provide detailed information during the execution
of a function or operation. By calling \code{verbose_on}, you can enable
this additional output.

The \code{options("Rapi_verbose" = TRUE)} line sets the verbose option to
\code{TRUE}, allowing functions to produce more detailed messages.
}
\examples{
verbose_on()

}
\seealso{
\code{\link{verbose_off}}: Turn off verbose mode.
}
