% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekday_dist.R
\name{weekday_dist}
\alias{weekday_dist}
\title{Home office distribution over weekdays}
\usage{
weekday_dist(
  wfh,
  weekday = NULL,
  emp = c("full_time", "mult_part_time", "part_time")
)
}
\arguments{
\item{wfh}{Home office frequency (numeric)}

\item{weekday}{In English and lowercase (optional)}

\item{emp}{One of \code{c("full_time", "mult_part_time", "part_time")}}
}
\value{
\code{data.frame}
}
\description{
Conditional on the home office frequency as well as full-time, multi part-time or part-time
}
