% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAddressInfoByLocation.R
\name{getAddressInfoByLocation}
\alias{getAddressInfoByLocation}
\title{getAddressInfoByLocation}
\usage{
getAddressInfoByLocation(location, key)
}
\arguments{
\item{location}{Character class of location address}
}
\value{
A data.frame contains address, latitude and longitude of given location.
}
\description{
Returns latitude and longitude of the given input location.
}
\examples{
getAddressInfoByLocation("Mjölby")
getAddressInfoByLocation("Damascus University")
}
